+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * CORE: [FIX] Move node into the same parent does not end in an exception #271
 * CORE: [FIX] clone 0byte file ends in exception #259
 * CORE: [FIX] Force removal of a filtered collection ends in timeout or loss of nodes #272
+* CORE: [FIX] Move share/reference to sub folder of a share #267
 
 
 ## 2.4.1