+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.4.2
 **Maintainer**: Raffael Sahli <sahli@gyselroth.com>\
-**Date**: 
+**Date**: wed Feb 06 15:40:03 CET 2019 
 
 * CORE: [FIX] Move node from share to share ends in Balloon\\Filesystem\\Exception\\Conflict a node called a does already exists in this collection #268
 * CORE: [FIX] Omit parent filtered collection if itself part of the result set #263